requirements.txt
# Core data / ML
numpy
pandas
scipy
scikit-learn
joblib

# Plotting
matplotlib

# Deep learning (only needed if LSTM path is used)
torch
torchvision
torchaudio

# Gateway API prototype
fastapi
uvicorn
python-multipart
kafka-python
